pyproject.toml
[project]
name = "100daysofpython-day32-birthdaywisher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
]
